+++ packages/client/src/api/client.ts
@@ -1,5 +1,6 @@
 import { getAPIKey } from '../util/apiKey';
 import { getFetchImplementation } from '../util/fetch';
+import { isString } from '../util/lang';
 import type * as Types from './components';
 import { operationsByTag } from './components';
 import type { FetcherExtraProps, FetchImpl } from './fetcher';
@@ -235,6 +236,52 @@ class DatabaseApi {
       ...this.extraProps
     });
   }
+
+  public getGitBranchesMapping(
+    workspace: Schemas.WorkspaceID,
+    dbName: Schemas.DBName
+  ): Promise<Schemas.ListGitBranchesResponse> {
+    return operationsByTag.database.getGitBranchesMapping({
+      pathParams: { workspace, dbName },
+      ...this.extraProps
+    });
+  }
+
+  public addGitBranchesEntry(
+    workspace: Schemas.WorkspaceID,
+    dbName: Schemas.DBName,
+    body: Types.AddGitBranchesEntryRequestBody
+  ): Promise<Types.AddGitBranchesEntryResponse> {
+    return operationsByTag.database.addGitBranchesEntry({
+      pathParams: { workspace, dbName },
+      body,
+      ...this.extraProps
+    });
+  }
+
+  public removeGitBranchesEntry(
+    workspace: Schemas.WorkspaceID,
+    dbName: Schemas.DBName,
+    gitBranch: string
+  ): Promise<void> {
+    return operationsByTag.database.removeGitBranchesEntry({
+      pathParams: { workspace, dbName },
+      queryParams: { gitBranch },
+      ...this.extraProps
+    });
+  }
+
+  public resolveBranch(
+    workspace: Schemas.WorkspaceID,
+    dbName: Schemas.DBName,
+    gitBranch: string
+  ): Promise<Types.ResolveBranchResponse> {
+    return operationsByTag.database.resolveBranch({
+      pathParams: { workspace, dbName },
+      queryParams: { gitBranch },
+      ...this.extraProps
+    });
+  }
 }
 
 class BranchApi {
@@ -262,12 +309,12 @@ class BranchApi {
     workspace: Schemas.WorkspaceID,
     database: Schemas.DBName,
     branch: Schemas.BranchName,
-    from = '',
+    from?: string,
     options: Types.CreateBranchRequestBody = {}
   ): Promise<void> {
     return operationsByTag.branch.createBranch({
       pathParams: { workspace, dbBranchName: `${database}:${branch}` },
-      queryParams: { from },
+      queryParams: isString(from) ? { from } : undefined,
       body: options,
       ...this.extraProps
     });