+++ cli/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @xata.io/cli
 
+## 0.1.3
+
+### Patch Changes
+
+- [#289](https://github.com/xataio/client-ts/pull/289) [`9a41035`](https://github.com/xataio/client-ts/commit/9a410354911d89116b332dd3ed7345b1156f74ef) Thanks [@gimenete](https://github.com/gimenete)! - Fix execution of the cli from npx
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`6d76275`](https://github.com/xataio/client-ts/commit/6d7627555a404a4c2da42f4187df6f8300f9a46f), [`d1ec0df`](https://github.com/xataio/client-ts/commit/d1ec0df14834088a816919bfc68216f3f9b2d9ef), [`1864742`](https://github.com/xataio/client-ts/commit/18647428d8608841de514c3784fb711c39dccc6d), [`1af6f1a`](https://github.com/xataio/client-ts/commit/1af6f1aaa1123e77a895961581c87f06a88db698), [`be4eda8`](https://github.com/xataio/client-ts/commit/be4eda8f73037d97fef7de28b56d7471dd867875), [`99be734`](https://github.com/xataio/client-ts/commit/99be734827576d888aa12a579ed1983a0a8a8e83)]:
+  - @xata.io/client@0.10.0
+  - @xata.io/codegen@0.10.0
+  - @xata.io/shell@0.1.4
+
+## 0.1.1
+
+### Patch Changes
+
+- [#273](https://github.com/xataio/client-ts/pull/273) [`30bfc10`](https://github.com/xataio/client-ts/commit/30bfc10f07ec0743cdaf15f1173e0487686a9d58) Thanks [@SferaDev](https://github.com/SferaDev)! - Add datetime columns
+
+## 0.1.0
+
+### Minor Changes
+
+- [#263](https://github.com/xataio/client-ts/pull/263) [`8ac045c`](https://github.com/xataio/client-ts/commit/8ac045c900f546f6abc2d4a73c38e07047b53583) Thanks [@gimenete](https://github.com/gimenete)! - Added command random-data
+
+### Patch Changes
+
+- [#258](https://github.com/xataio/client-ts/pull/258) [`8bd00a3`](https://github.com/xataio/client-ts/commit/8bd00a3efb2328d6700ee89ecdebc3356bc55d7f) Thanks [@gimenete](https://github.com/gimenete)! - Fixes a bug that make the CLI always generate TypeScript code
+
+* [#259](https://github.com/xataio/client-ts/pull/259) [`a2727a3`](https://github.com/xataio/client-ts/commit/a2727a3bf6dfddb41251ff86c40afe7851a9c068) Thanks [@gimenete](https://github.com/gimenete)! - Added command "schema edit"
+
 ## 0.0.1
 
 ### Patch Changes