+++ packages/codegen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xata.io/codegen
 
+## 0.10.0
+
+### Patch Changes
+
+- [#265](https://github.com/xataio/client-ts/pull/265) [`99be734`](https://github.com/xataio/client-ts/commit/99be734827576d888aa12a579ed1983a0a8a8e83) Thanks [@SferaDev](https://github.com/SferaDev)! - Add datetime field type support
+
+## 0.9.2
+
+### Patch Changes
+
+- [#273](https://github.com/xataio/client-ts/pull/273) [`30bfc10`](https://github.com/xataio/client-ts/commit/30bfc10f07ec0743cdaf15f1173e0487686a9d58) Thanks [@SferaDev](https://github.com/SferaDev)! - Add datetime columns
+
 ## 0.9.0
 
 ### Patch Changes